+++ tests/tests_pytorch/deprecated_api/test_remove_2-0.py
@@ -22,7 +22,6 @@ from torch.utils.data import DataLoader
 from pytorch_lightning.accelerators.cpu import CPUAccelerator
 from pytorch_lightning.core.mixins.device_dtype_mixin import DeviceDtypeModuleMixin
 from pytorch_lightning.demos.boring_classes import RandomDataset
-from pytorch_lightning.plugins.environments import LightningEnvironment
 from pytorch_lightning.strategies.utils import on_colab_kaggle
 from pytorch_lightning.utilities.apply_func import (
     apply_to_collection,
@@ -43,16 +42,6 @@ from pytorch_lightning.utilities.device_parser import (
     parse_gpu_ids,
     parse_hpus,
     parse_tpu_cores,
-)
-from pytorch_lightning.utilities.distributed import (
-    all_gather_ddp_if_available,
-    distributed_available,
-    gather_all_tensors,
-    get_default_process_group_backend_for_device,
-    init_dist_connection,
-    sync_ddp,
-    sync_ddp_if_available,
-    tpu_distributed,
 )
 from pytorch_lightning.utilities.optimizer import optimizer_to_device, optimizers_to_device
 from pytorch_lightning.utilities.seed import pl_worker_init_function, reset_seed, seed_everything
@@ -87,11 +76,6 @@ def test_v1_10_deprecated_xla_device_utilities():
 
     with pytest.deprecated_call(match="xla_device.XLADeviceUtils.tpu_device_exists` has been deprecated in v1.8.0"):
         XLADeviceUtils.tpu_device_exists()
-
-    from pytorch_lightning.utilities.distributed import tpu_distributed
-
-    with pytest.deprecated_call(match="tpu_distributed` has been deprecated in v1.8.0"):
-        tpu_distributed()
 
 
 def test_v1_10_deprecated_apply_func_utilities():
@@ -162,39 +146,6 @@ def test_v1_10_deprecated_device_parser_utilities():
         parse_tpu_cores(None)
 
 
-def test_v1_10_deprecated_distributed_utilities():
-    with pytest.deprecated_call(match="distributed.all_gather_ddp_if_available` has been deprecated in v1.8.0"):
-        all_gather_ddp_if_available(torch.tensor(1))
-
-    with pytest.deprecated_call(match="distributed.distributed_available` has been deprecated in v1.8.0"):
-        distributed_available()
-
-    with mock.patch("torch.distributed.get_world_size", return_value=2), mock.patch(
-        "torch.distributed.barrier"
-    ), mock.patch("torch.distributed.all_gather"):
-        with pytest.deprecated_call(match="distributed.gather_all_tensors` has been deprecated in v1.8.0"):
-            gather_all_tensors(torch.tensor(1))
-
-    with pytest.deprecated_call(
-        match="distributed.get_default_process_group_backend_for_device` has been deprecated in v1.8.0"
-    ):
-        get_default_process_group_backend_for_device(torch.device("cpu"))
-
-    with mock.patch("torch.distributed.is_initialized", return_value=True):
-        with pytest.deprecated_call(match="distributed.init_dist_connection` has been deprecated in v1.8.0"):
-            init_dist_connection(LightningEnvironment(), "gloo")
-
-    with pytest.deprecated_call(match="distributed.sync_ddp_if_available` has been deprecated in v1.8.0"):
-        sync_ddp_if_available(torch.tensor(1))
-
-    with mock.patch("torch.distributed.barrier"), mock.patch("torch.distributed.all_reduce"):
-        with pytest.deprecated_call(match="distributed.sync_ddp` has been deprecated in v1.8.0"):
-            sync_ddp(torch.tensor(1))
-
-    with pytest.deprecated_call(match="distributed.tpu_distributed` has been deprecated in v1.8.0"):
-        tpu_distributed()
-
-
 def test_v1_10_deprecated_optimizer_utilities():
     with pytest.deprecated_call(match="optimizer.optimizers_to_device` has been deprecated in v1.8.0"):
         optimizers_to_device([torch.optim.Adam(torch.nn.Linear(1, 1).parameters())], "cpu")
@@ -217,10 +168,3 @@ def test_v1_10_deprecated_seed_utilities():
 def test_v1_10_deprecated_accelerator_setup_environment_method():
     with pytest.deprecated_call(match="`Accelerator.setup_environment` has been deprecated in deprecated in v1.8.0"):
         CPUAccelerator().setup_environment(torch.device("cpu"))
-
-
-def test_v1_8_1_deprecated_rank_zero_only():
-    from pytorch_lightning.utilities.distributed import rank_zero_only
-
-    with pytest.deprecated_call(match="rank_zero_only` has been deprecated in v1.8.1"):
-        rank_zero_only(lambda: None)